[PR-1222] Write an article about parser combinators and QQ
Problem: We wish to have an article describing parser combinators. As a
bonus, we also add a section on how to use our parser with QQ.

Solution: Make an article describing how to make a simple parser
combinator from scratch, then how to use Megaparsec, and finally a bonus
section on how to integrate our S-expression parser into quasi
quotations.