[PR-1268] Fix mention of typed splices
Problem: I incorrectly said that in GHC < 9 you need to write `$$(foo)``
instead of `$$foo`. This is not true in all situations.

Solution: Remove this part. Instead put an introduction to this section.
Below, write a more detailed paragraph about this, mentioning that in
GHC 8 it works in some cases and in some it won't as opposed to GHC 9.