[PR-1268] Talk about untyped TH change in GHC 9
Problem: Seems like one change was made in GHC 9 for the untyped TH
code. `PlainTV` is not parametric over a flag.

Solution: Say that for GHC 9, one needs to write `TyVarBndr ()` instead,
and write `PlainTV _ ()`. Fix new article so it says GHC 9 is "mostly
unchaged" (it was the only required changed in that article).