[PR-1246] Make corrections
Problem: Non-negative integers under addition are a free monoid
(which is isomorphic to the list monoid described in the blog post,
given that the generator set S is a singleton).

Solution: Remove "non-negative", expand a little on what is meant
by "no additional laws are implied", and add a details section
explaining why there was an error and how removing "non-negative"
fixes it.