Rename `Elem`'s constructor to `MkElem`
Problem: Defining a type like `data D = D` introduces some ambiguity -
if we mention `D`, are we talking about the type `D` or the constructor
`D`?

In teaching materials, it's usually better to give types and
constructors different names to avoid this ambiguity.

Solution: Rename `Elem`'s constructor to `MkElem` to remove this
ambiguity.