Fix `SomeException` section
Problem: There are a few problems with this quote:

> As long as you haven't specified its type (and it is not obvious to
> the compiler, for example, from the handler), all exceptions will be
> caught, wrapped in `SomeException`.

First, it's not clear what "as long as you haven't specified its type"
means. After some head scratching, I realized it's referring to the
`catch` function, which is not mentioned anywhere else in the post. It
should be rephrased to something like:

> When you catch an exception with, for example, the `catch` function,
> and you don't tell it which exception type to catch, ...

Second, even if this was made clearer, it would still be wrong. The
`catch` function has a type variable `e`, which you MUST instantiate to
`SomeException` (either implicitly or explicitly) to catch all
exceptions.

It's not a matter of not specifying which exception you want to catch -
you MUST specify it and it must be `SomeException`.

Third, this is irrelevant and distracts from the point of the blog post.
These are all details of the `catch` function, but here we're focusing
on the `SomeException` type, let's keep our focus on that. If we want to
illustrate how to catch a `SomeException`, we should use `catchAll`
instead of `catch`.

----

This other quote is also misleading:

> However, again, there is a limitation. We know nothing about the inner
> exception.

We do actually know a few things about the inner exception: we know it
has an `Exception` instance, which implies a `Show` and a `Typeable`
instance.

----

> We can, for example, print it

If we're going to state that the exception can be printed, we should
probably be explicit about _why_ we have enough information to be able
to print it.

In other words, we should at least explain that we can print it because
we have an `Exception` instance in scope, and `Exception` implies
`Show`.

Solution: rewrite this bit, focus on the information we have in scope
when matching on the `SomeException` constructor and the information we
_don't_ have.