Rephrase paragraph about matching on `MkElem`
Problem:
> Since the type of elements is unknown at the use site, you can't
> address it

First: We're talking about a data type's constructor here. A constructor
can be used for two things: constructing a new value or deconstructing
via pattern matching. So when we say "at the use site", we have to be
specific: are we talking about creating a new value with `MkElem` or
pattern matching on `MkElem`?

Second: It's not clear what "addressing a type" means.

Solution: Rephrase this paragraph.