Edits
The section "how it works" has a lot of good stuff, but I feel it didn't quite work. I tried for quite some time to reorganize stuff to fix it. The problem was that derivation is opposite to parsing, so we can't really explain derivation terms with the help of constructs we use for parsing. 

I hope it's good now? I spent a lot of time and it's 4AM now. 🙈 Also contributed some edits above.